requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
websockets>=10.4
pyyaml>=6.0.1
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
pytest-cov>=4.1.0
